Draft 1
I have the landing page (pretty much how I want it, except for the
links that don’t work).

I started the design page to see how the thumbnail gallery will work,
and the close-up of a test project.

There is no Javascript yet.
The footer is missing.
The pages “More”, “About” and “Contact” are missing.